Add switchable KCL memory backend facade
